pyproject.toml
[project]
name = "tcc-llm-rag-rda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.50.1",
    "python-dotenv>=1.2.1",
]
